History 모델 생성
