+++ my_health/settings/base.py
@@ -44,6 +44,7 @@ INSTALLED_APPS = [
     # local apps
     "app_user",
     "app_workout",
+    "app_history",
 ]
 
 MIDDLEWARE = [
